dialog improvements
