+++ client/src/components/game/helpDialog/HelpDialog.css
